refactor: move archives debug_mode to respective folder
