+++ app/src/main/res/layout/activity_create_subject.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".view.activity.CreateSubjectActivity">
+    tools:context=".debug_mode.view.activity.CreateSubjectActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
