+++ app/src/main/res/layout/activity_select_option.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.activity.SelectOptionActivity">
+    tools:context=".debug_mode.view.activity.SelectOptionActivity">
 
     <include
         android:id="@+id/header_options"
